Improve roles and users backend tests
